Remove GEN logic from pipeline driver
